Refine code structure - Backend and Frontend
