+++ app/client/src/layout/header.css
@@ -1,0 +1,38 @@
+
+
+.header {
+    background-color: #036; /* BC Blue */
+    font-weight: 600;
+    color: white;
+     padding-top: 5px;
+    margin:auto;
+    width:100%;
+    display: inline-flex;
+    border-bottom: solid;
+    border-bottom-color: #e3a82b; /* BC Yellow*/
+    border-bottom-width: 0.4vh;
+    box-shadow: 0px 0px 6px grey;
+    
+  }
+
+  .App-logo {
+    padding-left: 5px;
+    height: 60px;
+    width: 200px;
+  }
+  #Profile{
+
+    position: absolute;
+    right: 40px;
+    height: 60px;
+    width: 60px;
+    padding-top: 2.5px;
+  }
+
+  #Menue{
+    position: absolute;
+    right: 5px;
+    height: 60px;
+    width: 60px;
+  }
+
