+++ app/client/src/layout/sideMenu.css
@@ -1,0 +1,74 @@
+/* 
+ Copyright © 2022 Province of British Columbia
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
+/**
+ * Application entry point
+ * @author [Jayna Bettesworth](bettesworthjayna@gmail.com)
+ * @module
+ */
+ .nav{
+    height: 100%;
+    width: 0;
+    margin-right: 0px;
+    right: 0;
+    position: fixed;
+    z-index: 1;
+    top:72px;
+    background-color: #036;
+    padding-top: 0px;
+    transition: 0.7s;
+    overflow-x: hidden;
+    opacity: 0.9;
+  }
+  
+  .nav ul{
+    padding: 0;
+    padding-top: 25px;
+  }
+  
+  .nav li:hover{
+    border-left: 7px solid #e3a82b;
+    background-color: rgb(2, 69, 136);
+  }
+  
+  .nav li {
+    display: block;
+    padding: 10px 20px;
+    color: #ccc;
+    text-decoration: none;
+    font-size: 15px;
+  }
+  
+  .nav .close {
+    position: absolute;
+    top: 0;
+    right: 20px;
+    font-size: 30px;;
+  }
+  
+  .nav a {
+    display: block;
+    padding: 2px;
+    color: #ccc;
+    text-decoration: none;
+    font-size: 25px;
+  }
+  #Menu{
+    height: 60px;
+    width: 60px;
+    position: absolute;
+    right: 5px;
+  }
