+++ app/client/src/layout/sideMenu.js
@@ -21,9 +21,7 @@
  */
  import React, { Component } from 'react';
  import {Link} from 'react-router-dom';
- import './header.css';
- import BCLogo from './icons/BCLogo.svg';
- import UserPic from './icons/user.png';
+ import './sideMenu.css';
  import Menu from './icons/menuLogo.svg';
 
  const Header = () => {
@@ -41,37 +39,30 @@
   }
 
       return (
-        <div className='header' >
+        <div id='content'>
+            <span className='slide'>
+              <a href="#" onClick={openSlideMenu}>
+                <img src={Menu} id='Menu' alt="Profile" /> 
+              </a> 
+            </span>
 
-            
-            <img src={BCLogo} className="App-logo" alt="logo" />
-            <h2> The Neighbourhood </h2>
-            
-            <div id='content'>
-              <span className='slide'>
-                <a href="#" onClick={openSlideMenu}>
-                  <img src={Menu} id='Menu' alt="Profile" /> 
-                </a> 
-              </span>
-
-              <div id='menu' className='nav'>
-                <a href="#" className='close' onClick={closeSlideMenu} >
+            <div id='menu' className='nav'>
+             <a href="#" className='close' onClick={closeSlideMenu} >
                 ✖
-                </a>
-                <ul>
-                  <li> <Link to='/dashboard' >Dashboard</Link></li>
-                  <li><Link to='./about' >About</Link></li>
-                  <li><Link to='./communities' >Communities</Link></li>
-                  <li><Link to='./profile/:id' >Profile</Link></li>
-                  <li><Link to='./login' >Log Off</Link></li>
-                </ul>
-                
-
-              </div>
+             </a>
+            <ul>
+                <li> <Link to='/dashboard' >Dashboard</Link></li>
+                <li><Link to='./about' >About</Link></li>
+                <li><Link to='./communities' >Communities</Link></li>
+                <li><Link to='./profile/:id' >Profile</Link></li>
+                <li><Link to='./login' >Log Off</Link></li>
+            </ul>
+          
 
             </div>
-            
-        </div>
+
+         </div>
+
       )
     }
