+++ app/server/index.js
@@ -26,190 +26,27 @@
  const port = process.env.PORT || 5000;
  const cors = require('cors')
  const mongoose = require('mongoose')
- const User = require('./models/user.model')
- const Community = require('./models/community.model')
- const jwt = require('jsonwebtoken')
- const bcrypt = require('bcryptjs') //encrypting passwords
+
+ 
+ const registerRouter = require('./routes/register')
+ const loginRouter = require('./routes/login')
+ const profileRouter = require('./routes/profile')
+ const editProfileRouter = require('./routes/editProfile')
+ const communitiesRouter = require('./routes/communities')
  
   app.use(cors()) //middleware
   app.use(express.json())
  
 
 
-  mongoose.connect('mongodb://localhost:27017/TheNeighborhood')
+mongoose.connect('mongodb://localhost:27017/TheNeighborhood')
+
+ app.use("/api/register", registerRouter);
+ app.use('/api/login', loginRouter);
+ app.use('/api/profile', profileRouter);
+ app.use('/api/editProfile', editProfileRouter);
+ app.use('/api/community', communitiesRouter);
  
- app.post('/api/register', async (req, res) => {
-     console.log(req.body)
-     try {
-         
-         const newPassword = await bcrypt.hash(req.body.password, 10)
-         
-         await User.create ({
-             name: req.body.name,
-             email: req.body.email,
-             password: newPassword,
-         })
-        
-         res.json({ status: 'ok' })
-     } catch (err) {
-        
-         res.json({ status: 'error', error: 'Duplicate email' })
-         
-     }
- })
-
-
- 
- app.post('/api/login', async (req, res) => {
-     const user = await User.findOne({
-         name: req.body.name,
-     })
- 
-     if (!user) {
-         return { status: 'error', error: 'Invalid login' }
-     }
- 
-     const isPasswordValid = await bcrypt.compare(
-         req.body.password,
-         user.password
-     )
- 
-     if (isPasswordValid) {
-         const token = jwt.sign(
-             {
-                 name: user.name,
-                 email: user.email,
-             },
-             'secret123'
-         )
- 
-         return res.json({ status: 'ok', user: token })
-     } else {
-         return res.json({ status: 'error', user: false })
-     }
- })
- 
-
- app.get('/api/quote', async (req, res) => {
-     const token = req.headers['x-access-token']
- 
-     try {
-         const decoded = jwt.verify(token, 'secret123')
-         const email = decoded.email
-         const user = await User.findOne({ email: email })
- 
-         return res.json({ status: 'ok', name: user.name })
-     } catch (error) {
-         console.log(error)
-         res.json({ status: 'error', error: 'invalid token' })
-     }
- })
- 
-
- app.get('/api/profile', async (req, res) => {
-    const token = req.headers['x-access-token']
-
-    try {
-        const decoded = jwt.verify(token, 'secret123')
-        const email = decoded.email
-        const user = await User.findOne({ email: email })
-
-        return res.json({ status: 'ok', name: user.name, email: user.email, fullName: user.fullName, bio: user.bio, title: user.title, quote:user.quote })
-    } catch (error) {
-        console.log(error)
-        res.json({ status: 'error', error: 'invalid token' })
-    }
-})
-
-app.post('/api/profile', async (req, res) => {
-    const token = req.headers['x-access-token']
-
-    try {
-        const decoded = jwt.verify(token, 'secret123')
-        const email = decoded.email
-        await User.updateOne(
-            { email: email },
-            { $set: { fullName: req.body.fullName } },
-        )
-
-        return res.json({ status: 'ok' })
-    } catch (error) {
-        console.log(error)
-        res.json({ status: 'error', error: 'invalid token' })
-    }
-})
-
-app.get('/api/editprofile', async (req, res) => {
-    const token = req.headers['x-access-token']
-
-    try {
-        const decoded = jwt.verify(token, 'secret123')
-        const email = decoded.email
-        const user = await User.findOne({ email: email })
-
-        return res.json({ 
-            status: 'ok', 
-            name: user.name, 
-            email: user.email, 
-            fullName: user.fullName, 
-            bio: user.bio, 
-            title: user.title, 
-            quote:user.quote 
-        })
-    } catch (error) {
-        console.log(error)
-        res.json({ status: 'error', error: 'invalid token' })
-    }
-})
-
-app.post('/api/editprofile', async (req, res) => {
-    const token = req.headers['x-access-token']
-
-    try {
-        const decoded = jwt.verify(token, 'secret123')
-        const email = decoded.email
-        await User.updateMany(
-            { email: email },
-            { $set: { fullName: req.body.fullName, title: req.body.title, bio: req.body.bio} },
-            
-        )
-
-        return res.json({ status: 'ok' })
-    } catch (error) {
-        console.log(error)
-        res.json({ status: 'error', error: 'invalid token' })
-    }
-})
-
-app.post('/api/Community', async (req, res) => {
-    console.log(req.body)
-    try {
-        
-        
-        await Community.create ({
-            title: req.body.title,
-            description: req.body.description,
-        })
-       
-        res.json({ status: 'ok' })
-    } catch (err) {
-       
-        res.json({ status: 'error', error: err })
-    }
-})
-
-app.get('/api/Community', async (req, res) => {
-    Community.find({ })
-        .then((data)=>{
-            console.log(data);
-            res.json(data);
-        })
-        .catch((error) =>{
-            console.log(error);
-        });
-
-})
-
 
  app.listen(port, () => {
      console.log(`Server started on port ${port}`)
