+++ app/server/routes/editProfile.js
@@ -1,0 +1,72 @@
+/* 
+ Copyright © 2022 Province of British Columbia
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
+/**
+ * Application entry point
+ * @author [Jayna Bettesworth](bettesworthjayna@gmail.com)
+ * @module
+ */
+
+ const express = require('express');
+ const router = express.Router();
+
+ const User = require('../models/user.model')
+ const jwt = require('jsonwebtoken')
+
+ 
+router.get('/', async (req, res) => {
+    const token = req.headers['x-access-token']
+
+    try {
+        const decoded = jwt.verify(token, 'secret123')
+        const email = decoded.email
+        const user = await User.findOne({ email: email })
+
+        return res.json({ 
+            status: 'ok', 
+            name: user.name, 
+            email: user.email, 
+            fullName: user.fullName, 
+            bio: user.bio, 
+            title: user.title, 
+            quote:user.quote 
+        })
+    } catch (error) {
+        console.log(error)
+        res.json({ status: 'error', error: 'invalid token' })
+    }
+})
+
+router.post('/', async (req, res) => {
+    const token = req.headers['x-access-token']
+
+    try {
+        const decoded = jwt.verify(token, 'secret123')
+        const email = decoded.email
+        await User.updateMany(
+            { email: email },
+            { $set: { fullName: req.body.fullName, title: req.body.title, bio: req.body.bio} },
+            
+        )
+
+        return res.json({ status: 'ok' })
+    } catch (error) {
+        console.log(error)
+        res.json({ status: 'error', error: 'invalid token' })
+    }
+})
+
+ module.exports = router;
