+++ app/server/routes/login.js
@@ -1,0 +1,60 @@
+/* 
+ Copyright © 2022 Province of British Columbia
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ */
+
+/**
+ * Application entry point
+ * @author [Jayna Bettesworth](bettesworthjayna@gmail.com)
+ * @module
+ */
+
+ const express = require('express');
+ const router = express.Router();
+
+ const User = require('../models/user.model')
+ const bcrypt = require('bcryptjs') //encrypting passwords
+ const jwt = require('jsonwebtoken')
+ 
+
+router.post('/', async (req, res) => {
+    const user = await User.findOne({
+        name: req.body.name,
+    })
+
+    if (!user) {
+        return { status: 'error', error: 'Invalid login' }
+    }
+
+    const isPasswordValid = await bcrypt.compare(
+        req.body.password,
+        user.password
+    )
+
+    if (isPasswordValid) {
+        const token = jwt.sign(
+            {
+                name: user.name,
+                email: user.email,
+            },
+            'secret123'
+        )
+
+        return res.json({ status: 'ok', user: token })
+    } else {
+        return res.json({ status: 'error', user: false })
+    }
+})
+
+module.exports = router;
